Fixed .socks remaining after the program finished running

+++ php/src/Services/Util/Asynchronous/AMPParallelAsynchronousProcessor.php
@@ -16,8 +16,11 @@ use Kinikit\Core\Asynchronous\Processor\AsynchronousProcessor;
 use Kinikit\Core\DependencyInjection\Container;
 use Kinikit\Core\Logging\Logger;
 use Kinikit\Core\Reflection\ClassInspectorProvider;
+use function Amp\async;
+use function Amp\delay;
 use function Amp\Future\await;
 use function Amp\Future\awaitAll;
+use function Amp\Future\awaitFirst;
 use function Amp\Parallel\Worker\submit;
 use function Amp\Parallel\Worker\workerPool;
 
@@ -58,8 +61,14 @@ class AMPParallelAsynchronousProcessor implements AsynchronousProcessor {
         $nInstances = count($asynchronousInstances);
         Logger::log("Parallel Processor Errors: " . $nFailed . " || Parallel Processor AsyncInstances " . $nInstances);
 
-        $workerPool->kill();
-        $workerPool->shutdown();
+        awaitFirst([
+            async(fn() => $workerPool->shutdown()),
+            async(fn() => delay(10, false))
+        ]);
+        if ($workerPool->isRunning()){
+            $workerPool->kill();
+            $workerPool->shutdown();
+        }
         unset($workerPool);
 
         return $asynchronousInstances;
